+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/display/FreBitmapDataKotlin.kt
@@ -30,6 +30,33 @@ import com.tuarua.frekotlin.FreException
 
 import java.nio.ByteBuffer
 
+/**
+ * A wrapper for handling FREBitmapData
+ *
+ * @param value the FREBitmapData to create this class from.
+ * @param bitmap the Bitmap to create this class from.
+ * @param swapColors swap the colors.
+ *
+ * @property rawValue the raw FREBitmapData value
+ * @property width An Int that specifies the width, in pixels, of the bitmap. This value corresponds to the
+ * width property of the ActionScript BitmapData class object. This field is read-only.
+ * @property height An Int that specifies the height, in pixels, of the bitmap. This value corresponds to the height
+ * property of the ActionScript BitmapData class object. This field is read-only.
+ * @property hasAlpha A Bool that indicates whether the bitmap supports per-pixel transparency. This value corresponds
+ * to the transparent property of the ActionScript BitmapData class object. If the value is non-zero, then the
+ * pixel format is ARGB32. If the value is zero, the pixel format is _RGB32. Whether the value is big endian
+ * or little endian depends on the host device. This field is read-only.
+ * @property isPremultiplied A Bool that indicates whether the bitmap pixels are stored as premultiplied color
+ * values. A true value means the values are premultipled. This field is read-only.
+ * @property lineStride32 An Int that specifies the number of UInt values per scanline. This value is typically
+ * the same as the width parameter. This field is read-only.
+ * @property isInvertedY A Bool that indicates the order in which the rows of bitmap data in the image are stored. A
+ * non-zero value means that the bottom row of the image appears first in the image data (in other words, the first
+ * value in the bits32 array is the first pixel of the last row in the image). A zero value means that the top row of
+ * the image appears first in the image data. This field is read-only.
+ * @property bits32 A ByteBuffer of the bitmap.
+ * @constructor
+ */
 class FreBitmapDataKotlin {
     var rawValue: FREBitmapData? = null
     var width = 0
@@ -40,28 +67,47 @@ class FreBitmapDataKotlin {
     var isInvertedY: Boolean = false
     var bits32: ByteBuffer? = null
 
-    constructor() {}
+    /**
+     *
+     *
+     */
+    constructor()
 
+    /**
+     *
+     *
+     */
     constructor(value: FreBitmapDataKotlin) {
         rawValue = value.rawValue
     }
 
+    /**
+     *
+     *
+     */
     constructor(value: FREObject) {
         rawValue = value as FREBitmapData
     }
 
+    /**
+     *
+     *
+     */
     constructor(value: FREBitmapData) {
         rawValue = value
     }
 
+    /**
+     *
+     *
+     */
     @Throws(FREASErrorException::class, FREWrongThreadException::class, FREInvalidObjectException::class)
     constructor(bitmap: Bitmap, swapColors: Boolean = true) {
         val fillColor = arrayOf<Byte>(0, 0, 0, 0)
         rawValue = FREBitmapData.newBitmapData(bitmap.width,
                 bitmap.height, bitmap.hasAlpha(), fillColor)
 
-        val bmp: Bitmap?
-        bmp = when {
+        val bmp: Bitmap = when {
             swapColors -> Companion.doSwapColors(bitmap)
             else -> bitmap
         }
@@ -72,6 +118,13 @@ class FreBitmapDataKotlin {
         bmp.recycle()
     }
 
+    /**
+     * See the original [Adobe documentation](https://help.adobe.com/en_US/air/extensions/WS11d1def534ea1be0-67da1176132a8869db6-7ffe.html)
+     *
+     * @throws
+     * @exception FREWrongThreadException
+     *
+     */
     @Throws(FREWrongThreadException::class, FREInvalidObjectException::class)
     fun acquire() {
         if (rawValue is FREBitmapData) {
@@ -86,11 +139,25 @@ class FreBitmapDataKotlin {
         }
     }
 
+    /**
+     * See the original [Adobe documentation](https://help.adobe.com/en_US/air/extensions/WS11d1def534ea1be0-67da1176132a8869db6-7ffe.html)
+     *
+     * @throws
+     * @exception FREWrongThreadException
+     * @exception FREInvalidObjectException
+     */
     @Throws(FREWrongThreadException::class, FREInvalidObjectException::class)
     fun release() {
         rawValue?.release()
     }
 
+    /**
+     * See the original [Adobe documentation](https://help.adobe.com/en_US/air/extensions/WS11d1def534ea1be0-67da1176132a8869db6-7ffe.html)
+     *
+     * @throws
+     * @exception FREWrongThreadException
+     * @exception FREInvalidObjectException
+     */
     @Throws(FREWrongThreadException::class, FREInvalidObjectException::class)
     fun invalidateRect(x: Int, y: Int, width: Int, height: Int) {
         rawValue?.invalidateRect(x, y, width, height)
@@ -111,7 +178,11 @@ class FreBitmapDataKotlin {
     }
 
 }
-
+/**
+ * converts a FREObject of type BitmapData to a Bitmap
+ */
+@Suppress("FunctionName")
+@JvmOverloads
 @Throws(FreException::class)
 fun Bitmap(freObject: FREObject?, swapColors: Boolean = true): Bitmap? {
     var ret: Bitmap? = null
@@ -136,4 +207,15 @@ fun Bitmap(freObject: FREObject?, swapColors: Boolean = true): Bitmap? {
         }
     }
     return ret
+}
+
+/**
+ * converts a Bitmap to a FREObject of type BitmapData
+ */
+fun Bitmap.toFREObject(): FREObject? {
+    return try {
+        FreBitmapDataKotlin(this,true).rawValue
+    } catch (e: Exception) {
+        FreException(e, "cannot create new object from Bitmap").getError(Thread.currentThread().stackTrace)
+    }
 }