+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/geom/FrePointKotlin.kt
@@ -13,6 +13,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+@file:Suppress("FunctionName", "unused")
+
 package com.tuarua.frekotlin.geom
 
 import android.util.Log