+++ starter_project/native_extension/ane/build-android.sh
@@ -6,10 +6,7 @@ pathtome=$0
 pathtome="${pathtome%/*}"
 AIR_SDK="/Users/User/sdks/AIR/AIRSDK_28"
 
-sh ./frekotlin-build.sh
-
-PROJECTNAME=FreKotlinExampleANE
-ANENAME=com.tuarua.frekotlin.example
+PROJECTNAME=HelloWorldANE
 
 #Copy SWC into place.
 echo "Copying SWC into place."
@@ -27,21 +24,29 @@ echo "Copying Android aars into place"
 cp "$pathtome/../../native_library/android/$PROJECTNAME/app/build/outputs/aar/app-release.aar" "$pathtome/platforms/android/app-release.aar"
 echo "getting Android jars"
 unzip "$pathtome/platforms/android/app-release.aar" "classes.jar" -d "$pathtome/platforms/android"
-
+unzip "$pathtome/platforms/android/app-release.aar" "res/*" -d "$pathtome/platforms/android"
+mv "$pathtome/platforms/android/res" "$pathtome/platforms/android/com.mycompany.$PROJECTNAME-res"
 
 #Run the build command.
 echo "Building ANE."
 "$AIR_SDK"/bin/adt -package \
--target ane "$pathtome/$ANENAME.ane" "$pathtome/extension.xml" \
+-target ane "$pathtome/$PROJECTNAME.ane" "$pathtome/extension_android.xml" \
 -swc "$pathtome/$PROJECTNAME.swc" \
 -platform Android-ARM \
 -C "$pathtome/platforms/android" "library.swf" "classes.jar" \
-com.tuarua.$PROJECTNAME-res/. \
+com.mycompany.$PROJECTNAME-res/. \
+-platformoptions "$pathtome/platforms/android/platform.xml" \
+-platform Android-x86 \
+-C "$pathtome/platforms/android" "library.swf" "classes.jar" \
+com.mycompany.$PROJECTNAME-res/. \
 -platformoptions "$pathtome/platforms/android/platform.xml" \
 
 rm "$pathtome/platforms/android/classes.jar"
 rm "$pathtome/platforms/android/app-release.aar"
 rm "$pathtome/platforms/android/library.swf"
+
 rm "$pathtome/$PROJECTNAME.swc"
 rm "$pathtome/library.swf"
+rm -r "$pathtome/platforms/android/com.mycompany.$PROJECTNAME-res"
+
 echo "DONE!"